refactor: dirname
