+++ src/utils/datetime.py
@@ -22,3 +22,7 @@ def now_dirname():
 def now_timestamp():
     format_string = DATETIME_BASE_FORMAT
     return datetime.datetime.now().strftime(format_string)
+
+
+NOW_DIRNAME = now_dirname()
+NOW_TIMESTAMP = now_timestamp()
